fix UB
Bench: 4266172

+++ src/search.cpp
@@ -580,7 +580,7 @@ moves_loop:
 		{
 			//calculate by how much we should reduce the search depth 
 			depth_reduction = reduction(pv_node, improving, depth, moves_searched);
-			depth_reduction -= std::min(3LL, GetHistoryScore(pos, sd, move, ss) / 16000);
+			depth_reduction -= std::min(static_cast<int64_t>(3), GetHistoryScore(pos, sd, move, ss) / 16000);
 			//adjust the reduction so that we can't drop into Qsearch and to prevent extensions
 			depth_reduction = std::min(depth - 1, std::max(depth_reduction, 1));
 			// search current move with reduced depth:
